Splitting project up into a package of modules for clarity
