+++ convert_to_irssi_logs.py
@@ -1,447 +1,14 @@
 import asyncio
-import datetime
 import json
-import os
 import sys
-from typing import Optional, List, Set, Dict
+from typing import Dict
 
-import sqlalchemy
 import telethon
-from telethon.tl.functions.messages import GetHistoryRequest
-from telethon.tl.types import MessageActionChatDeleteUser, MessageActionChatAddUser, MessageMediaDocument, \
-    MessageMediaPhoto, InputPeerChannel
-from telethon.tl.types.messages import Messages
-from tqdm import tqdm
+from telethon.tl.types import InputPeerChannel
 
-
-class Database:
-    def __init__(self, db_str: str) -> None:
-        self.engine = sqlalchemy.create_engine(db_str)
-        self.conn = self.engine.connect()
-        self.metadata = sqlalchemy.MetaData()
-        self.chat_logs = sqlalchemy.Table(
-            "telepisg_chat_logs",
-            self.metadata,
-            sqlalchemy.Column("chat_handle", sqlalchemy.String(), nullable=False, primary_key=True),
-            sqlalchemy.Column("last_message_id", sqlalchemy.Integer())
-        )
-        self.log_entries = sqlalchemy.Table(
-            "telepisg_log_entries",
-            self.metadata,
-            sqlalchemy.Column(
-                "chat_handle",
-                sqlalchemy.String(),
-                sqlalchemy.ForeignKey(
-                    "telepisg_chat_logs.chat_handle",
-                    ondelete="CASCADE"
-                ),
-                nullable=False
-            ),
-            sqlalchemy.Column("datetime", sqlalchemy.DateTime()),
-            sqlalchemy.Column("entry_type", sqlalchemy.String()),
-            sqlalchemy.Column("user_id", sqlalchemy.Integer()),
-            sqlalchemy.Column("message_id", sqlalchemy.Integer()),
-            sqlalchemy.Column("sub_message_id", sqlalchemy.Integer()),
-            sqlalchemy.Column("text", sqlalchemy.Text()),
-            sqlalchemy.UniqueConstraint("chat_handle", "message_id", "sub_message_id")
-        )
-        self.metadata.create_all(self.engine)
-
-    def insert_log_entries(self, chat_handle: str, log_entries: List["LogEntry"]):
-        query = sqlalchemy.insert(self.log_entries)
-        values_list = [
-            log_entry.to_row(chat_handle) for log_entry in log_entries
-        ]
-        self.conn.execute(query, values_list)
-
-    def list_log_dates(self, chat_handle: str) -> List[datetime.date]:
-        cols = [sqlalchemy.cast(self.log_entries.columns.datetime, sqlalchemy.Date)]
-        if self.engine.url.drivername == "sqlite":
-            cols = [sqlalchemy.func.DATE(self.log_entries.columns.datetime).label("datetime")]
-        query = sqlalchemy.select(
-            cols
-        ).distinct(
-        ).where(
-            self.log_entries.columns.chat_handle == chat_handle
-        ).order_by(
-            sqlalchemy.asc(self.log_entries.columns.datetime)
-        )
-        result = self.conn.execute(query)
-        rows = result.fetchall()
-        if self.engine.url.drivername == "sqlite":
-            return [
-                datetime.date.fromisoformat(row.datetime) for row in rows
-            ]
-        return [
-            row.datetime for row in rows
-        ]
-
-    def list_log_entries(self, chat_handle: str, log_date: Optional[datetime.date]):
-        conditions = [
-            self.log_entries.columns.chat_handle == chat_handle
-        ]
-        if log_date:
-            start_datetime = datetime.datetime.combine(log_date, datetime.time(0, 0, 0))
-            end_datetime = start_datetime + datetime.timedelta(days=1)
-            conditions.extend([
-                self.log_entries.columns.datetime >= start_datetime,
-                self.log_entries.columns.datetime < end_datetime
-            ])
-        query = sqlalchemy.select(
-            self.log_entries.columns
-        ).where(
-            sqlalchemy.and_(*conditions)
-        ).order_by(
-            sqlalchemy.asc(self.log_entries.columns.message_id),
-            sqlalchemy.asc(self.log_entries.columns.sub_message_id)
-        )
-        result = self.conn.execute(query)
-        return [
-            LogEntry.from_row(row)
-            for row in result.fetchall()
-        ]
-
-    def list_user_ids(self, chat_handle: str) -> Set[int]:
-        query = sqlalchemy.select(
-            self.log_entries.columns.user_id
-        ).distinct(
-            self.log_entries.columns.user_id
-        ).where(
-            self.log_entries.columns.chat_handle == chat_handle
-        )
-        result = self.conn.execute(query)
-        return set([row.user_id for row in result.fetchall()])
-
-    def update_chat_log(self, chat_handle: str, last_message_id: Optional[int]):
-        query = sqlalchemy.update(
-            self.chat_logs
-        ).values(
-            last_message_id=last_message_id
-        ).where(
-            self.chat_logs.columns.chat_handle == chat_handle
-        )
-        self.conn.execute(query)
-
-    def create_chat_log(self, chat_handle: str) -> None:
-        query = sqlalchemy.insert(
-            self.chat_logs
-        ).values(
-            chat_handle=chat_handle,
-            last_message_id=None
-        )
-        self.conn.execute(query)
-
-    def get_chat_log(self, chat_handle: str) -> "ChatLog":
-        query = sqlalchemy.select(
-            self.chat_logs.columns.last_message_id
-        ).where(
-            self.chat_logs.columns.chat_handle == chat_handle
-        )
-        result = self.conn.execute(query)
-        row = result.fetchone()
-        if row is None:
-            self.create_chat_log(chat_handle)
-            return ChatLog(chat_handle, self)
-        return ChatLog(chat_handle, self, last_message_id=row.last_message_id)
-
-
-class DataStore:
-
-    def __init__(self, db: Database, chat_handles: Optional[List] = None, user_ids: Optional[Set] = None):
-        self.db = db
-        self.chat_handles = chat_handles or []
-        self.user_ids = user_ids or set()
-        self.chat_logs = [ChatLog.load_from_database(chat_handle, self.db) for chat_handle in self.chat_handles]
-        self.user_extra_data = {}
-
-    def add_chat(self, chat_handle):
-        self.chat_handles.append(chat_handle)
-        self.chat_logs.append(ChatLog.load_from_database(chat_handle, self.db))
-
-    def save_to_json(self):
-        os.makedirs("irclogs_cache", exist_ok=True)
-        with open("irclogs_cache/data_store.json", "w", encoding="utf-8") as f:
-            json.dump({
-                "chat_handles": self.chat_handles,
-                "user_ids": list(self.user_ids),
-                "user_extra_data": self.user_extra_data
-            }, f, indent=2)
-
-    @classmethod
-    def load_from_json(cls, db: Database):
-        try:
-            with open("irclogs_cache/data_store.json", "r", encoding="utf-8") as f:
-                data = json.load(f)
-            handles = data.get("chat_handles")
-            user_ids = set(data.get("user_ids", []))
-            user_extra_data = data.get("user_extra_data", {})
-        except FileNotFoundError:
-            handles = []
-            user_ids = set()
-            user_extra_data = {}
-        data_store = cls(db, handles, user_ids)
-        data_store.user_extra_data = user_extra_data
-        return data_store
-
-    async def update_all_logs(self, client):
-        for chat_log in self.chat_logs:
-            await chat_log.scrape_messages(client)
-            for user_id in self.db.list_user_ids(chat_log.handle):
-                self.user_ids.add(user_id)
-
-    async def write_all_logs(self, client):
-        user_id_lookup = {
-            user_id: get_user_name_unique_deleted(await client.get_entity(user_id))
-            for user_id in self.user_ids
-        }
-        for chat_log in tqdm(self.chat_logs):
-            chat_entity = await client.get_entity(chat_log.handle)
-            chat_name = get_chat_name(chat_entity)
-            chat_log.write_log_files(user_id_lookup, chat_name)
-
-    async def write_users_cfg(self, client):
-        users_cfg = []
-        os.makedirs("pisg_output/user_pics/", exist_ok=True)
-        deleted_account_count = 0
-        for user_id in tqdm(self.user_ids):
-            user_name = get_user_name(await client.get_entity(user_id))
-            pic = await client.download_profile_photo(user_id, f"pisg_output/user_pics/{user_id}.png")
-            user_data = self.user_extra_data.get(str(user_id), {})
-            if user_name == "DELETED_ACCOUNT":
-                deleted_account_count += 1
-                if "alias" not in user_data:
-                    user_data["alias"] = get_user_name_unique_deleted(await client.get_entity(user_id))
-                if "nick" not in user_data:
-                    user_data["nick"] = f"{user_name}{deleted_account_count}"
-            else:
-                if "nick" not in user_data:
-                    user_data["nick"] = user_name
-            if "pic" not in user_data and pic is not None:
-                user_data["pic"] = f"user_pics/{user_id}.png"
-            user_line = "<user " + " ".join(f"{key}=\"{value}\"" for key, value in user_data.items()) + ">"
-            users_cfg.append(user_line)
-        with open("users.cfg", "w", encoding="utf-8") as f:
-            f.write("\n".join(users_cfg))
-
-    async def write_channel_cfg(self, client):
-        # Write channel config
-        chats_cfg = []
-        for chat_handle in self.chat_handles:
-            chat_name = get_chat_name(await client.get_entity(chat_handle))
-            clean_name = chat_name.replace(" ", r"\ ")
-            chats_cfg.append(f"""
-        <channel="{chat_name}">
-             Logfile = "irclogs/*/{clean_name}*.log"
-             OutputFile = "pisg_output/{chat_name}.html"
-        </channel>""")
-        with open("chats.cfg", "w", encoding="utf-8") as f:
-            f.write("\n".join(chats_cfg))
-
-
-def get_chat_name(entity):
-    if hasattr(entity, "title"):
-        return f"#{entity.title}"
-    else:
-        return get_user_name(entity) or str(entity.id)
-
-
-async def get_message_count(client, entity, latest_id=0):
-    get_history = GetHistoryRequest(
-        peer=entity,
-        offset_id=0,
-        offset_date=None,
-        add_offset=0,
-        limit=1,
-        max_id=0,
-        min_id=latest_id or 0,
-        hash=0
-    )
-    history = await client(get_history)
-    if isinstance(history, Messages):
-        count = len(history.messages)
-    else:
-        count = history.count
-    return count
-
-
-class ChatLog:
-
-    def __init__(self, handle: str, db: Database, last_message_id: Optional[int] = None):
-        self.handle = handle
-        self.db = db
-        self.last_message_id = last_message_id
-
-    def add_entries(self, log_entries: List["LogEntry"]):
-        if log_entries is None:
-            return
-        self.db.insert_log_entries(self.handle, log_entries)
-
-    async def scrape_messages(self, client):
-        entity = await client.get_entity(self.handle)
-        count = await get_message_count(client, entity, self.last_message_id)
-        chat_name = get_chat_name(entity)
-        latest_id = None
-        print(f"- Updating {chat_name} logs")
-        with tqdm(total=count) as bar:
-            async for message in client.iter_messages(entity):
-                if latest_id is None:
-                    latest_id = message.id
-                if self.last_message_id is not None and message.id <= self.last_message_id:
-                    print(f"- Caught up on {chat_name}")
-                    break
-                else:
-                    self.add_entries(LogEntry.entries_from_message(message, chat_name))
-                bar.update(1)
-        self.last_message_id = latest_id
-        self.db.update_chat_log(self.handle, self.last_message_id)
-
-    @classmethod
-    def load_from_database(cls, chat_handle: str, database: Database) -> "ChatLog":
-        return database.get_chat_log(chat_handle)
-
-    def write_log_files(self, user_id_lookup, chat_name):
-        for log_date in self.db.list_log_dates(self.handle):
-            file_contents = [
-                "--- Log opened " + log_date.strftime("%a %b %d 00:00:00 %Y"),
-                *[entry.to_log_line(user_id_lookup) for entry in self.db.list_log_entries(self.handle, log_date)]
-            ]
-            if log_date != datetime.date.today():
-                next_date = log_date + datetime.timedelta(days=1)
-                file_contents.append("--- Log closed " + next_date.strftime("%a %b %d 00:00:00 %Y"))
-            os.makedirs(f"irclogs/{log_date.year}", exist_ok=True)
-            file_name = get_file_name(chat_name, log_date)
-            with open(file_name, "w", encoding="utf-8") as f:
-                f.write("\n".join(file_contents))
-
-
-class LogEntry:
-    TYPE_TEXT = "TEXT"
-    TYPE_JOIN = "JOIN"
-    TYPE_QUIT = "QUIT"
-    TYPE_ACTION = "ACTION"
-
-    def __init__(
-            self,
-            log_datetime: datetime.datetime,
-            log_type: str,
-            user_id: int,
-            text: str,
-            message_id: int,
-            sub_message_id: int
-    ):
-        self.log_datetime = log_datetime
-        self.log_type = log_type
-        self.user_id = user_id
-        self.text = text
-        self.message_id = message_id
-        self.sub_message_id = sub_message_id
-
-    def to_row(self, chat_handle: str) -> Dict:
-        return {
-            "chat_handle": chat_handle,
-            "datetime": self.log_datetime,
-            "entry_type": self.log_type,
-            "user_id": self.user_id,
-            "text": self.text,
-            "message_id": self.message_id,
-            "sub_message_id": self.sub_message_id
-        }
-
-    @staticmethod
-    def from_row(row) -> "LogEntry":
-        return LogEntry(
-            row.datetime,
-            row.entry_type,
-            row.user_id,
-            row.text,
-            row.message_id,
-            row.sub_message_id
-        )
-
-    @classmethod
-    def entries_from_message(cls, message, log_name):
-        if isinstance(message.action, MessageActionChatDeleteUser):
-            return [LogEntry(
-                message.date,
-                cls.TYPE_QUIT,
-                message.sender.id,
-                f"[~{message.sender.id}@Telegram] has quit [Left chat]",
-                message.id,
-                0
-            )]
-        elif isinstance(message.action, MessageActionChatAddUser):
-            return [LogEntry(
-                message.date,
-                cls.TYPE_JOIN,
-                message.sender.id,
-                f"[~{message.sender.id}@Telegram] has joined {log_name}",
-                message.id,
-                0
-            )]
-        elif message.text:
-            return [LogEntry(
-                message.date,
-                cls.TYPE_TEXT,
-                message.sender.id,
-                text,
-                message.id,
-                sub_id
-            ) for sub_id, text in enumerate(message.text.split("\n")[::-1])]
-        elif message.media and isinstance(message.media, MessageMediaDocument):
-            return [LogEntry(
-                message.date,
-                cls.TYPE_ACTION,
-                message.sender.id,
-                f"sent a document ID={message.media.document.id}",
-                message.id,
-                0
-            )]
-        elif message.media and isinstance(message.media, MessageMediaPhoto):
-            return [LogEntry(
-                message.date,
-                cls.TYPE_ACTION,
-                message.sender.id,
-                f"sent a photo ID={message.media.photo.id}",
-                message.id,
-                0
-            )]
-
-    def to_log_line(self, user_id_lookup):
-        time = self.log_datetime.time().isoformat()
-        user_name = user_id_lookup[self.user_id]
-        if self.log_type == self.TYPE_QUIT:
-            return f"{time} -!- {user_name} {self.text}"
-        elif self.log_type == self.TYPE_JOIN:
-            return f"{time} -!- {user_name} {self.text}"
-        elif self.log_type == self.TYPE_TEXT:
-            return f"{time} < {user_name}> {self.text}"
-        elif self.log_type == self.TYPE_ACTION:
-            return f"{time} * {user_name} {self.text}"
-
-
-def try_int(handle):
-    try:
-        return int(handle.strip())
-    except ValueError:
-        return handle.strip()
-
-
-def get_user_name(user):
-    full_name = (user.first_name or "") + ("" if user.last_name is None else " " + user.last_name)
-    if full_name == "":
-        return "DELETED_ACCOUNT"
-    return full_name.replace(" ", "_")
-
-
-def get_user_name_unique_deleted(user):
-    full_name = (user.first_name or "") + ("" if user.last_name is None else " " + user.last_name)
-    if full_name == "":
-        return f"DELETED_ACCOUNT{user.id}"
-    return full_name.replace(" ", "_")
-
-
-def get_file_name(log_name, log_date):
-    return f"irclogs/{log_date.year}/{log_name}.{log_date.strftime('%m-%d')}.log"
+from telegram_logger.data_store import DataStore
+from telegram_logger.database import Database
+from telegram_logger.telegram_utils import get_chat_name
 
 
 async def add_channel(data_store, client):
