+++ telegram_logger/telegram_utils.py
@@ -1,0 +1,42 @@
+from telethon.tl.functions.messages import GetHistoryRequest
+from telethon.tl.types.messages import Messages
+
+
+def get_chat_name(entity):
+    if hasattr(entity, "title"):
+        return f"#{entity.title}"
+    else:
+        return get_user_name(entity) or str(entity.id)
+
+
+def get_user_name(user):
+    full_name = (user.first_name or "") + ("" if user.last_name is None else " " + user.last_name)
+    if full_name == "":
+        return "DELETED_ACCOUNT"
+    return full_name.replace(" ", "_")
+
+
+async def get_message_count(client, entity, latest_id=0):
+    get_history = GetHistoryRequest(
+        peer=entity,
+        offset_id=0,
+        offset_date=None,
+        add_offset=0,
+        limit=1,
+        max_id=0,
+        min_id=latest_id or 0,
+        hash=0
+    )
+    history = await client(get_history)
+    if isinstance(history, Messages):
+        count = len(history.messages)
+    else:
+        count = history.count
+    return count
+
+
+def get_user_name_unique_deleted(user):
+    full_name = (user.first_name or "") + ("" if user.last_name is None else " " + user.last_name)
+    if full_name == "":
+        return f"DELETED_ACCOUNT{user.id}"
+    return full_name.replace(" ", "_")
